Move base classes to the Base folder and update use statements.
